Oh crap I forgot I need to push

+++ src/main/java/com/sammy/omnis/common/items/ModShovelItem.java
@@ -17,7 +17,7 @@ import net.minecraftforge.common.ToolType;
 
 public class ModShovelItem extends ShovelItem
 {
-    public ModShovelItem(IItemTier material, int damage, float speed, Properties properties)
+    public ModShovelItem(IItemTier material, float damage, float speed, Properties properties)
     {
         super(material, damage + 1.5f, speed - 3f, properties.maxDamage(material.getMaxUses()).addToolType(ToolType.SHOVEL, material.getHarvestLevel()));
     }
